feat: acctually make it work
The for loop that greps over the input descriptor multiple times
actually cannot do that!  It can only do it once, after which you would
need to lseek to the start of the descriptor, but you cannot do that on
input.

To make it work I just used an actual file to write/read to and fro.